requirements.txt
numpy==1.26.0
pypng==0.20220715.0
